Employee Story Done
As far as the Employee being able to see all users, see their respective accounts and make Transactions in those accounts that is all finished.

+++ src/main/java/dmacc/beans/Account.java
@@ -92,7 +92,7 @@ public class Account {
 	/**
 	 * @param iD the iD to set
 	 */
-	public void setID(int iD) {
+	public void setID(long iD) {
 		ID = iD;
 	}
 	/**
